generic/python-device-weather-forecast/requirements.txt
paho-mqtt>=1.2
xmltodict >=0.10.2
protobuf>=3.1.0
schedule>=0.4.2
requests>=2.13.0
dateutils>=0.6.6
